Versao atualizada de acordo com os requisitos que tem no github

+++ src/main/kotlin/com/tryplace/tryplace/service/ImovelService.kt
@@ -2,24 +2,93 @@ package com.tryplace.tryplace.service
 
 import com.tryplace.tryplace.dto.ImovelDto
 import com.tryplace.tryplace.dto.ImovelRequest
+import com.tryplace.tryplace.dto.ImovelVisitanteDto
 import com.tryplace.tryplace.exceptions.RecursoNaoEncontradoException
+import com.tryplace.tryplace.exceptions.RegraDeNegocioException
 import com.tryplace.tryplace.model.ImovelModel
+import com.tryplace.tryplace.model.UsuarioModel
 import com.tryplace.tryplace.repository.ImovelRepository
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
+import java.math.BigDecimal
 import java.util.UUID
+import kotlin.String
 
 @Service
 class ImovelService(
     private val repository: ImovelRepository
 ) {
-    fun criarImovel (request: ImovelRequest): ImovelDto {
+
+    fun buscarPorPrecoMinMax(precoMin: BigDecimal, precoMax: BigDecimal): List<ImovelDto> {
+        if (precoMin > precoMax) {
+            throw RegraDeNegocioException("O valor mínimo não pode ser maior que o valor maximo")
+        }
+
+        return repository.findByValorAluguelBetween(precoMin, precoMax).map { converterParaCadastradoDto(it) }
+    }
+
+    fun listarImovelTodos(): List<ImovelVisitanteDto> {
+        return repository.findAll().map { converterParaVisitanteDto(it) }
+    }
+
+    fun buscaPorPrecoMaxVisitante(precoMax: BigDecimal): List<ImovelVisitanteDto> {
+        return repository.findByValorAluguelLessThanEqual(precoMax).map { converterParaVisitanteDto(it) }
+    }
+
+    fun buscaPorTipoImovelVisitante(tipoImovel: String): List<ImovelVisitanteDto> {
+        return repository.findByTipoImovel(tipoImovel).map { converterParaVisitanteDto(it) }
+    }
+
+    fun buscarImovelCadastradoPorId(id: UUID): ImovelDto {
+        val imovel = repository.findById(id)
+            .orElseThrow{ RecursoNaoEncontradoException("Imovel com $id não encontrado") }
+
+        return converterParaCadastradoDto(imovel)
+    }
+
+
+    private fun converterParaVisitanteDto(im: ImovelModel) = ImovelVisitanteDto(
+        id = im.id!!,
+        imagemImovel = im.imagemImovel,
+        valorAluguel = im.valorAluguel,
+        bairroImovel = im.bairroImovel,
+        quantidadeQuarto = im.quantidadeQuarto,
+        quantidadeBanheiro = im.quantidadeBanheiro,
+        tipoImovel = im.tipoImovel
+    )
+
+    private fun converterParaCadastradoDto(im: ImovelModel) = ImovelDto(
+        id = im.id!!,
+        nomeImovel = im.nomeImovel,
+        imagemImovel = im.imagemImovel,
+        valorAluguel = im.valorAluguel,
+        avaliacaoImovel = im.avaliacaoImovel ,
+        descricaoImovel = im.descricaoImovel,
+        quantidadeBanheiro = im.quantidadeBanheiro,
+        quantidadeQuarto = im.quantidadeQuarto,
+        tipoImovel = im.tipoImovel,
+        telefoneLocador = im.dono.telefone,
+        wifi = im.wifi,
+        cafeDaManha = im.cafeDaManha,
+        ruaImovel = im.ruaImovel,
+        numeroImovel = im.numeroImovel,
+        bairroImovel = im.bairroImovel,
+        cidadeImovel = im.cidadeImovel,
+        estadoImovel = im.estadoImovel,
+        cepImovel = im.cepImovel
+    )
+
+    fun criarImovel (request: ImovelRequest, donoLogado: UsuarioModel): ImovelDto {
         val imovel = ImovelModel (
             nomeImovel = request.nomeImovel.trim(),
             imagemImovel = request.imagemImovel,
+            valorAluguel = request.valorAluguel,
             avaliacaoImovel = request.avaliacaoImovel,
             descricaoImovel = request.descricaoImovel,
+            quantidadeQuarto = request.quantidadeQuarto,
+            quantidadeBanheiro = request.quantidadeBanheiro,
+            tipoImovel = request.tipoImovel,
             wifi = request.wifi,
             cafeDaManha = request.cafeDaManha,
             ruaImovel = request.ruaImovel.trim(),
@@ -27,30 +96,38 @@ class ImovelService(
             bairroImovel = request.bairroImovel.trim(),
             cidadeImovel = request.cidadeImovel.trim(),
             estadoImovel = request.estadoImovel.trim(),
-            cepImovel = request.cepImovel.trim()
+            cepImovel = request.cepImovel.trim(),
+            dono = donoLogado
         )
 
         val savedImovel = repository.save(imovel)
 
-        return mapearParaDto(savedImovel)
+        return converterParaCadastradoDto(savedImovel)
 
     }
 
     fun listarImovel(paginacao: Pageable): Page<ImovelDto> {
         val imoveisPage = repository.findAll(paginacao)
 
-        return imoveisPage.map { model -> mapearParaDto(model) }
+        return imoveisPage.map { model -> converterParaCadastradoDto(model) }
     }
 
-    fun editarImovel(id: UUID, request: ImovelRequest): ImovelDto {
+    fun editarImovel(id: UUID, request: ImovelRequest, donoLogado: UsuarioModel): ImovelDto {
         val imovelExistente = repository.findById(id).orElseThrow {
             RecursoNaoEncontradoException("Imóvel com ID $id não encontrado para edição")
+        }
+        if (imovelExistente.dono.id != donoLogado.id) {
+            throw RegraDeNegocioException("Você não tem permissão para editar esse imóvel")
         }
 
         imovelExistente.nomeImovel = request.nomeImovel.trim()
         imovelExistente.imagemImovel = request.imagemImovel
+        imovelExistente.valorAluguel = request.valorAluguel
         imovelExistente.avaliacaoImovel = request.avaliacaoImovel
         imovelExistente.descricaoImovel = request.descricaoImovel
+        imovelExistente.quantidadeQuarto = request.quantidadeQuarto
+        imovelExistente.quantidadeBanheiro = request.quantidadeBanheiro
+        imovelExistente.tipoImovel = request.tipoImovel
         imovelExistente.wifi = request.wifi
         imovelExistente.cafeDaManha = request.cafeDaManha
         imovelExistente.ruaImovel = request.ruaImovel.trim()
@@ -62,33 +139,19 @@ class ImovelService(
 
         val imovelAtualizado = repository.save(imovelExistente)
 
-        return mapearParaDto(imovelAtualizado)
+        return converterParaCadastradoDto(imovelAtualizado)
     }
 
-    fun deleteImovel(id: UUID) {
+    fun deleteImovel(id: UUID, donoLogado: UsuarioModel) {
         val imovelExistente = repository.findById(id).orElseThrow {
             RecursoNaoEncontradoException("Imóvel com ID $id não encontrado para deletar")
+        }
+        if (imovelExistente.dono.id != donoLogado.id) {
+            throw RegraDeNegocioException("Você não tem permissão para deletar esse imóvel")
         }
         repository.delete(imovelExistente)
     }
 
-    private fun mapearParaDto(imovel: ImovelModel): ImovelDto {
-        return ImovelDto(
-            id = imovel.id!!,
-            nomeImovel = imovel.nomeImovel,
-            imagemImovel = imovel.imagemImovel,
-            avaliacaoImovel = imovel.avaliacaoImovel,
-            descricaoImovel = imovel.descricaoImovel,
-            wifi = imovel.wifi,
-            cafeDaManha = imovel.cafeDaManha,
-            ruaImovel = imovel.ruaImovel,
-            numeroImovel = imovel.numeroImovel,
-            bairroImovel = imovel.bairroImovel,
-            cidadeImovel = imovel.cidadeImovel,
-            estadoImovel = imovel.estadoImovel,
-            cepImovel = imovel.cepImovel
-        )
-    }
 
 
 }
